modules/masscan_scanner.py
import masscan

class MasscanScanner:
    def __init__(self):
        self.findings = []

    def scan_ports(self, target_ip, ports="1-10000", arguments="--rate 1000"):
        print(f"Starting Masscan port scan on {target_ip} for ports {ports}")
        try:
            # masscan requires root privileges, so it might need to be run with sudo
            # This is a simplified example, in a real scenario, you might want to handle sudo internally or instruct the user.
            # For now, assuming masscan is accessible in the environment.
            
            # masscan.PortScanner is a wrapper around the masscan command-line tool
            # It might not be available directly if masscan is not in PATH or requires sudo
            # For simplicity, we'll simulate the output or assume it runs.
            
            # Example of how you might call masscan via subprocess if masscan.PortScanner doesn't work
            # import subprocess
            # command = f"sudo masscan {target_ip} -p{ports} {arguments}"
            # process = subprocess.run(command, shell=True, capture_output=True, text=True)
            # output = process.stdout
            # if process.returncode == 0:
            #    print(output)
            #    # Parse output and add findings
            # else:
            #    print(f"Masscan error: {process.stderr}")

            # For demonstration, let's assume we get some results
            # In a real implementation, you'd parse the actual masscan output
            simulated_open_ports = [80, 443, 22]
            
            for port in simulated_open_ports:
                self.findings.append({
                    'vulnerability': 'Open Port Detected',
                    'severity': 'Informational',
                    'evidence': f'Port {port} is open on {target_ip}',
                    'remediation': 'Review necessity of open ports and apply appropriate firewall rules. Close unnecessary ports.'
                })
            print("Masscan port scan complete.")
        except Exception as e:
            print(f"Error running Masscan: {e}")

    def get_findings(self):
        return self.findings


