requirements.txt
Flask==3.1.1
flask-cors==6.0.1
requests==2.32.4
beautifulsoup4
python-masscan
python-nmap
python-Wappalyzer
aiohttp
dnspython

